[map-helper-functions] Create useApplyLegendToRow.ts hook
